+++ upgrade/upgrade-1.12.0.php
@@ -1,0 +1,14 @@
+<?php
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+function upgrade_module_1_12_0($module)
+{
+    return $module instanceof Wilden_manager
+        && WmProfilePermission::ensureSuperAdmin()
+        && $module->registerHook('actionOrderGridDefinitionModifier')
+        && $module->registerHook('actionOrderGridQueryBuilderModifier')
+        && $module->registerHook('displayBackOfficeHeader');
+}